+++ dns/tokenizer.py
@@ -485,7 +485,7 @@ class Tokenizer(object):
             raise dns.exception.SyntaxError('expecting an identifier')
         if not token.value.isdigit():
             raise dns.exception.SyntaxError('expecting an integer')
-        value = long(token.value)
+        value = int(token.value)
         if value < 0 or value > 4294967296L:
             raise dns.exception.SyntaxError('%d is not an unsigned 32-bit integer' % value)
         return value
